docs/diffusion_processes/dp_requirements.txt
wandb==0.14.0
hydra-core==1.2.0
matplotlib==3.4.0
scikit-learn==1.2.2
scipy==1.9.2
torch==2.0.1
torchsde==0.2.5